# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  internal_web_controllers_http.SwaggerCreateUserRequest:
    properties:
      email:
        type: string
      fullName:
        type: string
      password:
        minLength: 6
        type: string
      permissions:
        items:
          $ref: '#/definitions/photogallery_api_go_internal_application_dtos_request_security.RolePermissionSetItemDTO'
        type: array
      phone:
        type: string
      roleId:
        type: string
      username:
        type: string
    required:
    - fullName
    - password
    - permissions
    - roleId
    - username
    type: object
...
  internal_web_controllers_http.SwaggerErrorResponse:
    properties:
      code:
        type: string
      detail:
        type: string
      errors:
        items:
          type: string
        type: array
      instance:
        type: string
      status:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
...
  internal_web_controllers_http.SwaggerMeta:
    properties:
      total:
        type: integer
    type: object
  internal_web_controllers_http.SwaggerOKFlag:
    properties:
      ok:
        type: boolean
    type: object
  internal_web_controllers_http.SwaggerOKFlagResponse:
    properties:
      data:
        $ref: '#/definitions/internal_web_controllers_http.SwaggerOKFlag'
    type: object
  internal_web_controllers_http.SwaggerOverrideItemRequest:
    properties:
      mode:
        enum:
        - Grant
        - Revoke
        type: string
      permissionId:
        type: string
      subMenuId:
        type: string
    required:
    - mode
    - permissionId
    - subMenuId
    type: object
  internal_web_controllers_http.SwaggerOverrideListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/photogallery_api_go_internal_domain_entities.UserPermissionOverride'
        type: array
    type: object
  internal_web_controllers_http.SwaggerOverrideResponse:
    properties:
      data:
        $ref: '#/definitions/photogallery_api_go_internal_domain_entities.UserPermissionOverride'
    type: object
...
  internal_web_controllers_http.SwaggerReplaceOverridesRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/photogallery_api_go_internal_application_dtos_request_security.OverrideItemDTO'
        type: array
    type: object
...
  internal_web_controllers_http.SwaggerSetPasswordRequest:
    properties:
      newPassword:
        minLength: 6
        type: string
    required:
    - newPassword
    type: object
  internal_web_controllers_http.SwaggerSetPrimaryRoleRequest:
    properties:
      primaryRoleId:
        type: string
    required:
    - primaryRoleId
    type: object
...
  internal_web_controllers_http.SwaggerUpdateUserRequest:
    properties:
      email:
        type: string
      fullName:
        type: string
      isActive:
        type: boolean
      permissions:
        items:
          $ref: '#/definitions/photogallery_api_go_internal_application_dtos_request_security.RolePermissionSetItemDTO'
        type: array
      phone:
        type: string
      roleId:
        type: string
    required:
    - fullName
    - permissions
    - roleId
    type: object
...
  internal_web_controllers_http.SwaggerUserListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/photogallery_api_go_internal_application_dtos_responses_security.UserResponseDTO'
        type: array
      meta:
        $ref: '#/definitions/internal_web_controllers_http.SwaggerMeta'
    type: object
  internal_web_controllers_http.SwaggerUserResponse:
    properties:
      data:
        $ref: '#/definitions/photogallery_api_go_internal_application_dtos_responses_security.UserResponseDTO'
    type: object
  internal_web_controllers_http.SwaggerUserRolesResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/photogallery_api_go_internal_domain_entities.UserRole'
        type: array
    type: object
...
  photogallery_api_go_internal_application_dtos_request_security.OverrideItemDTO:
    properties:
      mode:
        enum:
        - Grant
        - Revoke
        type: string
      permissionId:
        type: string
      subMenuId:
        type: string
    required:
    - mode
    - permissionId
    - subMenuId
    type: object
  photogallery_api_go_internal_application_dtos_request_security.RolePermissionSetItemDTO:
    properties:
      menuId:
        type: string
      permissionIds:
        items:
          type: string
        type: array
      subMenuId:
        type: string
    required:
    - menuId
    - permissionIds
    - subMenuId
    type: object
...
  photogallery_api_go_internal_application_dtos_responses_security.UserResponseDTO:
    properties:
      createdAtUtc:
        type: string
      email:
        type: string
      fullName:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      phone:
        type: string
      updatedAtUtc:
        type: string
      username:
        type: string
    type: object
...
  photogallery_api_go_internal_domain_entities.UserPermissionOverride:
    properties:
      createdAtUtc:
        type: string
      id:
        type: string
      mode:
        type: string
      permissionID:
        type: string
      subMenuID:
        type: string
      updatedAtUtc:
        type: string
      userID:
        type: string
    type: object
  photogallery_api_go_internal_domain_entities.UserRole:
    properties:
      isPrimary:
        type: boolean
      roleID:
        type: string
      userID:
        type: string
    type: object
...
info:
  contact: {}
  description: API REST v1 Foto-tienda
  title: Photo Gallery API
  version: "1.0"
paths:
...
  /api/v1/users:
    get:
      description: Lista paginada de usuarios con búsqueda y orden opcional.
      parameters:
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 25
        description: Page size
        in: query
        name: pageSize
        type: integer
      - description: Search text
        in: query
        name: q
        type: string
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: Sort direction
        enum:
        - asc
        - desc
        in: query
        name: dir
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerUserListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: ListUsers
      tags:
      - security
    post:
      consumes:
      - application/json
      description: Crea un nuevo usuario.
      parameters:
      - description: Create user payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_web_controllers_http.SwaggerCreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: CreateUser
      tags:
      - security
  /api/v1/users/:id:
    get:
      description: Obtiene un usuario por identificador.
      parameters:
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: GetUser
      tags:
      - security
    put:
      consumes:
      - application/json
      description: Actualiza un usuario existente.
      parameters:
      - description: Update user payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_web_controllers_http.SwaggerUpdateUserRequest'
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: UpdateUser
      tags:
      - security
  /api/v1/users/:id/activate:
    patch:
      description: Activa un usuario.
      parameters:
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerOKFlagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: ActivateUser
      tags:
      - security
  /api/v1/users/:id/deactivate:
    patch:
      description: Desactiva un usuario.
      parameters:
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerOKFlagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: DeactivateUser
      tags:
      - security
  /api/v1/users/:id/overrides:
    get:
      description: Lista paginada de overrides de usuario con búsqueda y orden opcional.
      parameters:
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 25
        description: Page size
        in: query
        name: pageSize
        type: integer
      - description: Search text
        in: query
        name: q
        type: string
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: Sort direction
        enum:
        - asc
        - desc
        in: query
        name: dir
        type: string
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerOverrideListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: ListUserOverrides
      tags:
      - security
    post:
      consumes:
      - application/json
      description: Crea un override para un usuario.
      parameters:
      - description: Create override payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_web_controllers_http.SwaggerOverrideItemRequest'
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerOverrideResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: CreateUserOverride
      tags:
      - security
    put:
      consumes:
      - application/json
      description: Reemplaza en bloque los overrides de un usuario.
      parameters:
      - description: Replace overrides payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_web_controllers_http.SwaggerReplaceOverridesRequest'
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerOKFlagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: ReplaceUserOverrides
      tags:
      - security
  /api/v1/users/:id/overrides/:overrideId:
    delete:
      description: Elimina un override de usuario por identificador.
      parameters:
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Override ID (UUID)
        in: path
        name: overrideId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerOKFlagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: DeleteUserOverride
      tags:
      - security
  /api/v1/users/:id/password:
    patch:
      consumes:
      - application/json
      description: Actualiza la contraseña de un usuario.
      parameters:
      - description: Set password payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_web_controllers_http.SwaggerSetPasswordRequest'
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerOKFlagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: SetUserPassword
      tags:
      - security
  /api/v1/users/:id/roles:
    get:
      description: Obtiene los roles asignados a un usuario.
      parameters:
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerUserRolesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: GetUserRoles
      tags:
      - security
    put:
      consumes:
      - application/json
      description: Reemplaza los roles asignados a un usuario.
      parameters:
      - description: Set primary role payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_web_controllers_http.SwaggerSetPrimaryRoleRequest'
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerOKFlagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_web_controllers_http.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: SetUserRoles
      tags:
      - security
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"